@@ -155,7 +155,7 @@ export default {
 
     genUri (service) {
       let user = `${service.method}:${service.password}`
-      user = new Buffer(user).toString('base64')
+      user = Buffer.from(user).toString('base64')
       user = user.replace(/\+/g, '-').replace(/\//g, '_').replace(/=+$/, '')
       let mark = encodeURIComponent(service.name)
       return `ss://${user}@${service.server}:${service.port}#${mark}`
